requirements.txt
flask
requests
numpy
flask_cors
psycopg2-binary
sqlalchemy
geoalchemy2
shapely
opencv-python
beautifulsoup4==4.12.2
openai==0.28.0
python-dotenv==1.0.0
